+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add CiaynAgent to support models that do not have, or are not good at, agentic function calling.
+
 ## [0.10.3] - 2024-12-27
 
 - Fix logging on interrupt.